documentation + some cleaning
