+++ dsp/src/com/pema4/scalesynth/base/generators/Generator.java
@@ -1,0 +1,30 @@
+package com.pema4.scalesynth.base.generators;
+
+import com.pema4.scalesynth.base.Component;
+import com.pema4.scalesynth.base.processors.Processor;
+
+/**
+ * Represent a source of audio signal.
+ */
+public interface Generator extends Component {
+    /**
+     * Generates of the audio.
+     * Note that content of {@code outputs} can be overwritten (this behaviour depends on a subclass).
+     *
+     * @param outputs buffers to place generated audio into.
+     * @param n       how many samples to generate.
+     */
+    void generate(double[][] outputs, int n);
+
+    /**
+     * Chains this generator with processor.
+     * Returned generator outputs audio, generated by the generator and affected by the processor.
+     *
+     * @param processor processor to use.
+     * @return a new Generator instance.
+     */
+    default Generator then(Processor processor) {
+        return ChainedGenerator.of(this, processor);
+    }
+}
+
