+++ dsp/src/com/pema4/scalesynth/base/parameters/ParameterChangeListener.java
@@ -2,6 +2,7 @@ package com.pema4.scalesynth.base.parameters;
 
 /**
  * Functional interface, implemented by listeners of {@code Parameter} instance.
+ *
  * @param <T> value type of {@code Parameter}.
  */
 public interface ParameterChangeListener<T> {
