+++ dsp/src/com/pema4/scalesynth/dsp/generators/DualOscillator.java
@@ -51,6 +51,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the master oscillator pulse pulse width.
+     * <p>
      * This value is supposed to be in range [0.01, 0.99] (because at 0 and at 1 there is no sound).
      * <p>
      * <p>Pulse width is a relative length of the lower part of rectangle wave. Examples:</p>
@@ -68,6 +69,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the master oscillator saw-pulse mix.
+     * <p>
      * This value is supposed to be in range [0, +1].
      * <p>
      * At 0 oscillator oscillator outputs sawtooth wave, at 1 - rectangle wave.
@@ -82,6 +84,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the slave oscillator relative pitch (in octaves).
+     * <p>
      * This value is supposed to be in range [-3, +3].
      *
      * @param slaveOctave how many octaves the slave oscillator is higher than the master
@@ -93,6 +96,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the slave oscillator relative pitch (in semitones).
+     * <p>
      * This value is supposed to be in range [-12, +12].
      *
      * @param slaveSemi how many semitones the slave oscillator is higher than the master
@@ -104,6 +108,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the slave oscillator relative pitch (in cents).
+     * <p>
      * This value is supposed to be in range [-50, +50].
      *
      * @param slaveFine how many octaves the slave oscillator is higher than the master
@@ -115,6 +120,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the slave oscillator pulse pulse width.
+     * <p>
      * This value is supposed to be in range [0.01, 0.99] (because at 0 and at 1 there is no sound).
      * <p>
      * <p>Pulse width is a relative length of the lower part of rectangle wave. Examples:</p>
@@ -132,6 +138,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the slave oscillator saw-pulse mix.
+     * <p>
      * This value is supposed to be in range [0, +1].
      * <p>
      * At 0 oscillator oscillator outputs sawtooth wave, at 1 - rectangle wave.
@@ -160,6 +167,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the master oscillator amplitude.
+     * <p>
      * This value is supposed to be in range [0, +1].
      *
      * @param masterAmplitude a new noise amplitude of the master oscillator.
@@ -172,6 +180,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the slave oscillator amplitude.
+     * <p>
      * This value is supposed to be in range [0, 1].
      *
      * @param slaveAmplitude a new noise amplitude of the slave oscillator.
@@ -184,6 +193,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the noise generator amplitude.
+     * <p>
      * This value is supposed to be in range [0, +1].
      *
      * @param noiseAmplitude a new noise amplitude of the noise generator.
@@ -194,6 +204,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets a drift coefficient of the oscillators.
+     * <p>
      * This value is supposed to be in range [0, +1].
      * <p>
      * Drifting is a small fluctuation of oscillator's pitch that can be heard in analogue gear.
@@ -208,6 +219,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets a number of subvoices playing in unison for one note.
+     * <p>
      * This value is supposed to be in range [1, 8].
      * <p>
      * Each unison subvoice has slightly different pitch and panning.
@@ -222,6 +234,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the maximum pitch offset of unison voices.
+     * <p>
      * This value is supposed to be in range [0, 1].
      * <p>
      * This parameter controls how much not "in tune" are unison voices.
@@ -237,6 +250,7 @@ public class DualOscillator implements Generator {
 
     /**
      * Sets the maximum pitch offset of unison voices.
+     * <p>
      * This value is supposed to be in range [0, 1].
      * <p>
      * Unison voices playing at lower pitches are moved to the left channel, and vice versa.
@@ -274,7 +288,8 @@ public class DualOscillator implements Generator {
     }
 
     /**
-     * Performs the generation of the audio.
+     * Generates of the audio.
+     * Note that content of {@code outputs} is overwritten.
      *
      * @param outputs buffers to place generated audio into.
      * @param n       how many samples to generate.
@@ -390,249 +405,254 @@ public class DualOscillator implements Generator {
         for (var slave : slaves)
             slave.reset(Math.random(), Math.random());
     }
-}
-
-/**
- * The implementation of VA oscillator based on BLEPs.
- * <p>
- * In short: resulting audio is the sum of naive aliased waveform and some correcting signal.
- * Result is an approximation of bandlimited waveform (without aliasing).
- */
-class BlepOscillator {
-    private double phase;
-    private double masterPhase;
-    private double nextValue;
-    private int pulseStage;
-    private double freq;
-    private double mix;
-    private double pulseWidth;
-    private double normalizedMasterFreq;
-    private double currValue;
-    private double sampleRate;
-    private double normalizedFreq;
-    private double amplitude = 1;
-    private boolean syncEnabled;
-    private double masterFreq;
 
     /**
-     * Constructs a new instance of the MystranOscillator class.
+     * The implementation of VA oscillator based on BLEPs.
+     * <p>
+     * In short: resulting audio is the sum of naive aliased waveform and some correcting signal.
+     * Result is an approximation of bandlimited waveform (without aliasing).
      */
-    public BlepOscillator() {
-        reset(0, 0);
-    }
+    static class BlepOscillator {
+        private double phase;
+        private double masterPhase;
+        private double nextValue;
+        private int pulseStage;
+        private double freq;
+        private double mix;
+        private double pulseWidth;
+        private double normalizedMasterFreq;
+        private double currValue;
+        private double sampleRate;
+        private double normalizedFreq;
+        private double amplitude = 1;
+        private boolean syncEnabled;
+        private double masterFreq;
 
-    /**
-     * Returns a value of the 2-sample PolyBLEP function. This function is for the first sample.
-     *
-     * @param t an inter-sample time when upwards discontinuity happened.
-     * @return value of that PolyBLEP function (positive value).
-     */
-    private static double poly3blep0(double t) {
-        // these are just sanity checks
-        // correct code doesn't need them
-        if (t < 0) return 0;
-        if (t > 1) return 1;
+        /**
+         * Constructs a new instance of the MystranOscillator class.
+         */
+        public BlepOscillator() {
+            reset(0, 0);
+        }
 
-        double t2 = t * t;
-        return t * t2 - 0.5f * t2 * t2;
-    }
+        /**
+         * Returns a value of the 2-sample PolyBLEP function. This function is for the first sample.
+         *
+         * @param t an inter-sample time when upwards discontinuity happened.
+         * @return value of that PolyBLEP function (positive value).
+         */
+        private static double poly3blep0(double t) {
+            // these are just sanity checks
+            // correct code doesn't need them
+            if (t < 0) return 0;
+            if (t > 1) return 1;
 
-    /**
-     * Returns a value of the 2-sample PolyBLEP function. This function is for the second sample.
-     *
-     * @param t an inter-sample time when upwards discontinuity happened.
-     * @return value of that PolyBLEP function (negative value).
-     */
-    private static double poly3blep1(double t) {
-        return -poly3blep0(1 - t);
-    }
+            double t2 = t * t;
+            return t * t2 - 0.5f * t2 * t2;
+        }
 
-    /**
-     * Resets a state of this oscillator (use zero phase for the best results).
-     *
-     * @param phase a new phase of the slave oscillator.
-     * @param masterPhase a new phase of the master oscillator.
-     */
-    public void reset(double phase, double masterPhase) {
-        this.phase = phase;
-        this.masterPhase = masterPhase;
-        nextValue = 0;
-        pulseStage = 0;
-    }
+        /**
+         * Returns a value of the 2-sample PolyBLEP function. This function is for the second sample.
+         *
+         * @param t an inter-sample time when upwards discontinuity happened.
+         * @return value of that PolyBLEP function (negative value).
+         */
+        private static double poly3blep1(double t) {
+            return -poly3blep0(1 - t);
+        }
 
-    /**
-     * Generates a new audio and adds it to the buffer.
-     *
-     * @param output  buffer to add generated audio into.
-     * @param n       how many samples to generate.
-     */
-    public void generate(double[] output, int n) {
-        for (int i = 0; i < n; ++i) {
-            // take the delayed part from previous sample
-            currValue = nextValue;
-
-            // reset the delay so we can build into it
+        /**
+         * Resets a state of this oscillator (use zero phase for the best results).
+         *
+         * @param phase       a new phase of the slave oscillator.
+         * @param masterPhase a new phase of the master oscillator.
+         */
+        public void reset(double phase, double masterPhase) {
+            this.phase = phase;
+            this.masterPhase = masterPhase;
             nextValue = 0;
+            pulseStage = 0;
+        }
 
-            // then proceed like a trivial oscillator
-            phase += normalizedFreq;
-            masterPhase += normalizedMasterFreq;
+        /**
+         * Generates a new audio and adds it to the buffer.
+         *
+         * @param output buffer to add generated audio into.
+         * @param n      how many samples to generate.
+         */
+        public void generate(double[] output, int n) {
+            for (int i = 0; i < n; ++i) {
+                // take the delayed part from previous sample
+                currValue = nextValue;
 
-            // Then replace the reset logic: loop until we
-            // can't find the next discontinuity, so during
-            // one sample we can process any number of them!
-            while (true) {
-                // Now in order of the stages of the wave-form
-                // check for discontinuity during this sample.
+                // reset the delay so we can build into it
+                nextValue = 0;
 
-                // Very firstly, process master oscillator reset. This part is buggy.
-                if (syncEnabled && masterPhase > 1) {
-                    double exactResetTime = (masterPhase - 1) / normalizedMasterFreq;
-                    double exactSlavePhase = phase - normalizedFreq * (exactResetTime);
-                    double exactSlavePhaseCopy = exactSlavePhase;
-                    while (true) {
-                        if (pulseStage == 0) {
-                            if (exactSlavePhaseCopy < pulseWidth)
-                                break;
-                            handleUpwardsDiscontinuity(exactSlavePhaseCopy);
+                // then proceed like a trivial oscillator
+                phase += normalizedFreq;
+                masterPhase += normalizedMasterFreq;
+
+                // Then replace the reset logic: loop until we
+                // can't find the next discontinuity, so during
+                // one sample we can process any number of them!
+                while (true) {
+                    // Now in order of the stages of the wave-form
+                    // check for discontinuity during this sample.
+
+                    // Very firstly, process master oscillator reset. This part is buggy.
+                    if (syncEnabled && masterPhase > 1) {
+                        double exactResetTime = (masterPhase - 1) / normalizedMasterFreq;
+                        double exactSlavePhase = phase - normalizedFreq * (exactResetTime);
+                        double exactSlavePhaseCopy = exactSlavePhase;
+                        while (true) {
+                            if (pulseStage == 0) {
+                                if (exactSlavePhaseCopy < pulseWidth)
+                                    break;
+                                handleUpwardsDiscontinuity(exactSlavePhaseCopy);
+                            }
+
+                            if (pulseStage == 1) {
+                                if (exactSlavePhaseCopy < 1)
+                                    break;
+                                handleDownwardsDiscontinuity(exactSlavePhaseCopy);
+                                exactSlavePhaseCopy -= 1;
+                            }
                         }
 
-                        if (pulseStage == 1) {
-                            if (exactSlavePhaseCopy < 1)
-                                break;
-                            handleDownwardsDiscontinuity(exactSlavePhaseCopy);
-                            exactSlavePhaseCopy -= 1;
-                        }
+                        double blepAmp = ((1 - mix) * exactSlavePhaseCopy + mix * ((pulseStage == 1 ? 1 : 0)));
+
+                        currValue -= blepAmp * poly3blep0(exactResetTime);
+                        nextValue -= blepAmp * poly3blep1(exactResetTime);
+
+                        masterPhase -= 1;
+                        phase -= exactSlavePhase;
+                        pulseStage = 0;
                     }
 
-                    double blepAmp = ((1 - mix) * exactSlavePhaseCopy + mix * ((pulseStage == 1 ? 1 : 0)));
+                    // First is the upwards transition.
+                    if (pulseStage == 0) {
+                        if (phase < pulseWidth)
+                            break;
 
-                    currValue -= blepAmp * poly3blep0(exactResetTime);
-                    nextValue -= blepAmp * poly3blep1(exactResetTime);
+                        handleUpwardsDiscontinuity(phase);
+                    }
 
-                    masterPhase -= 1;
-                    phase -= exactSlavePhase;
-                    pulseStage = 0;
+                    // Second is the downwards transition.
+                    if (pulseStage == 1) {
+                        if (phase < 1)
+                            break;
+
+                        handleDownwardsDiscontinuity(phase);
+                        phase -= 1;
+                    }
                 }
 
-                // First is the upwards transition.
-                if (pulseStage == 0) {
-                    if (phase < pulseWidth)
-                        break;
+                // add naively generated value to the buffer.
+                nextValue += (1 - mix) * phase + mix * (pulseStage != 0 ? 1 : 0);
 
-                    handleUpwardsDiscontinuity(phase);
-                }
-
-                // Second is the downwards transition.
-                if (pulseStage == 1) {
-                    if (phase < 1)
-                        break;
-
-                    handleDownwardsDiscontinuity(phase);
-                    phase -= 1;
-                }
+                // and output is just what we collected, but
+                // let's make it range [-amplitude, amplitude]
+                output[i] += amplitude * (2 * currValue - 1);
             }
+        }
 
-            // add naively generated value to the buffer.
-            nextValue += (1 - mix) * phase + mix * (pulseStage != 0 ? 1 : 0);
+        private void handleDownwardsDiscontinuity(double slavePhase) {
+            // inter-sample time when transition happened.
+            double t = (slavePhase - 1) / normalizedFreq;
 
-            // and output is just what we collected, but
-            // let's make it range [-amplitude, amplitude]
-            output[i] += amplitude * (2 * currValue - 1);
+            // and negative transition.. normally you would
+            // calculate step-sizes for all mixed waves, but
+            // here both saw and pulse go from 1 to 0.. so
+            // it's always the same transition size!
+            currValue -= poly3blep0(t);
+            nextValue -= poly3blep1(t);
+
+            // and then we do a reset (just one!)
+            pulseStage = 0;
+        }
+
+        private void handleUpwardsDiscontinuity(double slavePhase) {
+            // inter-sample time when transition happened.
+            double t = (slavePhase - pulseWidth) / normalizedFreq;
+
+            // so then scale by pulse mix
+            // and add the first sample to output
+            currValue += mix * poly3blep0(t);
+
+            // and second sample to delay
+            nextValue += mix * poly3blep1(t);
+
+            // and then we proceed to next stage
+            pulseStage = 1;
+        }
+
+        /**
+         * Sets a sample rate used in processing.
+         *
+         * @param sampleRate new sample rate.
+         */
+        public void setSampleRate(double sampleRate) {
+            this.sampleRate = sampleRate;
+            normalizedFreq = freq / sampleRate;
+            normalizedMasterFreq = masterFreq / sampleRate;
+            reset(0, 0);
+        }
+
+        /**
+         * Sets a new value of this oscillator saw-pulse mix (described in {@link DualOscillator})
+         *
+         * @param mix a new value of saw-pulse mix.
+         */
+        public void setMix(double mix) {
+            this.mix = mix;
+        }
+
+        /**
+         * Sets a new value of this oscillator pulse width (described in {@link DualOscillator})
+         *
+         * @param pulseWidth a new value of pulse width.
+         */
+        public void setPulseWidth(double pulseWidth) {
+            this.pulseWidth = pulseWidth;
+        }
+
+        /**
+         * Sets a new value of slave oscillator's frequency.
+         *
+         * @param freq a new value of slave oscillator's frequency.
+         */
+        public void setFreq(double freq) {
+            this.freq = freq;
+            this.normalizedFreq = freq / sampleRate; // this algorithm works with normalized frequencies.
+        }
+
+        /**
+         * Sets a new value of master oscillator's frequency.
+         *
+         * @param masterFreq a new value of master oscillator's frequency.
+         */
+        public void setMasterFreq(double masterFreq) {
+            this.masterFreq = masterFreq;
+            this.normalizedMasterFreq = masterFreq / sampleRate;
+        }
+
+        /**
+         * Sets if the slave oscillator is hard synced to the master oscillator frequency.
+         *
+         * @param syncEnabled if the slave oscillator is hard synced to the master oscillator frequency.
+         */
+        public void setSyncEnabled(boolean syncEnabled) {
+            this.syncEnabled = syncEnabled;
+        }
+
+        /**
+         * Sets a new value of the output amplitude.
+         *
+         * @param amplitude a new value of the output amplitude.
+         */
+        public void setAmplitude(double amplitude) {
+            this.amplitude = amplitude;
         }
     }
-
-    private void handleDownwardsDiscontinuity(double slavePhase) {
-        // inter-sample time when transition happened.
-        double t = (slavePhase - 1) / normalizedFreq;
-
-        // and negative transition.. normally you would
-        // calculate step-sizes for all mixed waves, but
-        // here both saw and pulse go from 1 to 0.. so
-        // it's always the same transition size!
-        currValue -= poly3blep0(t);
-        nextValue -= poly3blep1(t);
-
-        // and then we do a reset (just one!)
-        pulseStage = 0;
-    }
-
-    private void handleUpwardsDiscontinuity(double slavePhase) {
-        // inter-sample time when transition happened.
-        double t = (slavePhase - pulseWidth) / normalizedFreq;
-
-        // so then scale by pulse mix
-        // and add the first sample to output
-        currValue += mix * poly3blep0(t);
-
-        // and second sample to delay
-        nextValue += mix * poly3blep1(t);
-
-        // and then we proceed to next stage
-        pulseStage = 1;
-    }
-
-    /**
-     * Sets a sample rate used in processing.
-     *
-     * @param sampleRate new sample rate.
-     */
-    public void setSampleRate(double sampleRate) {
-        this.sampleRate = sampleRate;
-        normalizedFreq = freq / sampleRate;
-        normalizedMasterFreq = masterFreq / sampleRate;
-        reset(0, 0);
-    }
-
-    /**
-     * Sets a new value of this oscillator saw-pulse mix (described in {@link DualOscillator})
-     * @param mix a new value of saw-pulse mix.
-     */
-    public void setMix(double mix) {
-        this.mix = mix;
-    }
-
-    /**
-     * Sets a new value of this oscillator pulse width (described in {@link DualOscillator})
-     * @param pulseWidth a new value of pulse width.
-     */
-    public void setPulseWidth(double pulseWidth) {
-        this.pulseWidth = pulseWidth;
-    }
-
-    /**
-     * Sets a new value of slave oscillator's frequency.
-     * @param freq a new value of slave oscillator's frequency.
-     */
-    public void setFreq(double freq) {
-        this.freq = freq;
-        this.normalizedFreq = freq / sampleRate; // this algorithm works with normalized frequencies.
-    }
-
-    /**
-     * Sets a new value of master oscillator's frequency.
-     * @param masterFreq a new value of master oscillator's frequency.
-     */
-    public void setMasterFreq(double masterFreq) {
-        this.masterFreq = masterFreq;
-        this.normalizedMasterFreq = masterFreq / sampleRate;
-    }
-
-    /**
-     * Sets if the slave oscillator is hard synced to the master oscillator frequency.
-     *
-     * @param syncEnabled if the slave oscillator is hard synced to the master oscillator frequency.
-     */
-    public void setSyncEnabled(boolean syncEnabled) {
-        this.syncEnabled = syncEnabled;
-    }
-
-    /**
-     * Sets a new value of the output amplitude.
-     *
-     * @param amplitude a new value of the output amplitude.
-     */
-    public void setAmplitude(double amplitude) {
-        this.amplitude = amplitude;
-    }
 }
+
