+++ dsp/src/com/pema4/scalesynth/dsp/generators/Envelope.java
@@ -1,9 +1,23 @@
-package com.pema4.scalesynth.dsp.processors;
+package com.pema4.scalesynth.dsp.generators;
 
 import com.pema4.scalesynth.base.KeyboardEvent;
 import com.pema4.scalesynth.base.KeyboardEventType;
 import com.pema4.scalesynth.base.generators.Generator;
 
+/**
+ * Represents an ADSR-envelope - main modulation source in this synthesizer.
+ * <p>
+ * It has 4 stages:
+ * <p>
+ * 1. In attack stage output rises from 0 to 1. This stage is started with Note On message;
+ * <p>
+ * 2. In decay stage output decreases from 1 to sustain level;
+ * <p>
+ * 3. At sustain stage output is contant;
+ * <p>
+ * 4. At release stage outputs decreases from current envelope level to 0.
+ * This stage is started with Note Off message.
+ */
 public class Envelope implements Generator {
     private double sampleRate = 1; // чтобы не делить на 0
     private double attackCoef;
@@ -16,25 +30,60 @@ public class Envelope implements Generator {
     private double decayRate;
     private double releaseRate;
 
+    /**
+     * Sets a rate of the attack stage. Lower rate - faster attack stage.
+     * <p>
+     * This value is supposed to be in range [5, 100].
+     *
+     * @param attackRate a new attack rate.
+     */
     public synchronized void setAttackRate(double attackRate) {
         this.attackRate = attackRate;
         attackCoef = Math.pow(1 - 0.01 * attackRate * attackRate / 10000, 44100 / sampleRate);
     }
 
+    /**
+     * Sets a rate of the decay stage. Lower rate - faster decay stage.
+     * <p>
+     * This value is supposed to be in range [5, 100].
+     *
+     * @param decayRate a new decay rate.
+     */
     public synchronized void setDecayRate(double decayRate) {
         this.decayRate = decayRate;
         decayCoef = Math.pow(1 - 0.001 * decayRate / 100, 44100 / sampleRate);
     }
 
+    /**
+     * Sets a level of the sustain stage.
+     * <p>
+     * This value is supposed to be in range [0, 1].
+     *
+     * @param sustainLevel a new sustain level.
+     */
     public synchronized void setSustainLevel(double sustainLevel) {
         this.sustainLevel = sustainLevel;
     }
 
+    /**
+     * Sets a rate of the release stage. Lower rate - faster release stage.
+     * <p>
+     * This value is supposed to be in range [5, 100].
+     *
+     * @param releaseRate a new release rate.
+     */
     public synchronized void setReleaseRate(double releaseRate) {
         this.releaseRate = releaseRate;
         releaseCoef = Math.pow(1 - 0.001 * releaseRate / 100, 44100 / sampleRate);
     }
 
+    /**
+     * Generates of the audio.
+     * Note that content of {@code outputs} is overwritten.
+     *
+     * @param outputs buffers to place generated audio into.
+     * @param n       how many samples to generate.
+     */
     @Override
     public synchronized void generate(double[][] outputs, int n) {
         var output = outputs[0];
@@ -63,7 +112,7 @@ public class Envelope implements Generator {
                 case RELEASE:
                     value *= releaseCoef;
                     output[i] = value;
-                    if (closeToZero(value)) {
+                    if (value > -0.00000001 && value < 0.00000001) {
                         state = State.DISABLED;
                         value = 0;
                     }
@@ -72,10 +121,6 @@ public class Envelope implements Generator {
                     output[i] = 0;
                     break;
             }
-    }
-
-    private boolean closeToZero(double value) {
-        return value > -0.00000001 && value < 0.00000001;
     }
 
     /**
@@ -118,6 +163,9 @@ public class Envelope implements Generator {
         return state != State.DISABLED;
     }
 
+    /**
+     * Represents current state of the envelope.
+     */
     private enum State {
         ATTACK,
         DECAY,
