--- dsp/src/main/java/com/pema4/scalesynth/base/processors/Processor.java
@@ -1,7 +1,0 @@
-package com.pema4.scalesynth.base.processors;
-
-import com.pema4.scalesynth.base.Component;
-
-public interface Processor extends Component {
-    void process(double[][] inputs, int n);
-}
