+++ gui/src/com/pema4/scalesynth/gui/controls/LinearParameterTransform.java
