+++ gui/src/com/pema4/scalesynth/gui/models/SynthMidiAdapter.java
