+++ gui/src/com/pema4/scalesynth/gui/services/ScaleService.java
