+++ gui/src/com/pema4/scalesynth/gui/views/SlaveOscillatorEditorView.java
